posix: add block_array.h header
