+++ src/libpmemfile-posix/block_array.h
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2017, Intel Corporation
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *     * Redistributions of source code must retain the above copyright
+ *       notice, this list of conditions and the following disclaimer.
+ *
+ *     * Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in
+ *       the documentation and/or other materials provided with the
+ *       distribution.
+ *
+ *     * Neither the name of the copyright holder nor the names of its
+ *       contributors may be used to endorse or promote products derived
+ *       from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#ifndef PMEMFILE_BLOCK_ARRAY_H
+#define PMEMFILE_BLOCK_ARRAY_H
+
+struct pmemfile_block;
+struct pmemfile_vinode;
+
+/*
+ * block_list_insert_after -- allocates new block metadata,
+ * and inserts it into the linked list of
+ * blocks. Does not allocate block->data.
+ *
+ * Returns the address of the new block metadata.
+ */
+struct pmemfile_block *
+block_list_insert_after(struct pmemfile_vinode *vinode,
+			struct pmemfile_block *prev);
+
+/*
+ * Removes the block from the linked list of blocks. Deallocates
+ * block->data if set, and deallocates the block metadata.
+ *
+ * Returns a pointer to the preceding block metadata - the one that
+ * was the previous block before deallocating.
+ */
+struct pmemfile_block *
+block_list_remove(struct pmemfile_vinode *vinode,
+		struct pmemfile_block *block);
+
+#endif
